moved porchfile function away from kpt pkg
