+++ internal/kpt/pkg/pkg.go
@@ -24,7 +24,6 @@ import (
 	"sort"
 
 	rgfilev1alpha1 "github.com/GoogleContainerTools/kpt/pkg/api/resourcegroup/v1alpha1"
-	"github.com/nephio-project/porch/api/porch/v1alpha1"
 	"github.com/nephio-project/porch/internal/kpt/errors"
 	"github.com/nephio-project/porch/internal/kpt/types"
 	"github.com/nephio-project/porch/internal/kpt/util/git"
@@ -265,21 +264,6 @@ func DecodeKptfile(in io.Reader) (*kptfilev1.KptFile, error) {
 		return kf, err
 	}
 	return kf, nil
-}
-
-// decoder for Porchfile's in porch packages
-func DecodePorchfile(in io.Reader) (*v1alpha1.PorchFile, error) {
-	pf := &v1alpha1.PorchFile{}
-	c, err := io.ReadAll(in)
-	if err != nil {
-		return pf, err
-	}
-	d := yaml.NewDecoder(bytes.NewBuffer(c))
-	d.KnownFields(true)
-	if err := d.Decode(pf); err != nil {
-		return pf, err
-	}
-	return pf, nil
 }
 
 // CheckKptfileVersion verifies the apiVersion and kind of the resource
